Modifies error handler to extract exception cause when present.
Context:
When an error is raised while rendering a template, the error gets
wrapped in an ActionView::Template::Error. The error can be accessed
with the #cause method on Template::Error (or #original_exception for
older versions of rails).

Rails 5 introduces a change in how exceptions are handled. Exceptions
are no longer unwrapped, and the expectation is now that a rescue_from
handler will be specified for the exception.

oop_rails_server already relies on rescue_from to capture exceptions and
serialize them to json. Fortitude specs use this behavior to check that
certain error conditions lead to correct exceptions. Those fortitude
specs expect to see the inner exception though, not the wrapper
exception, so, the oop_rails_server ActionController exception handler
will now unwrap captured exception if possible.

https://github.com/rails/rails/commit/e35b98e6f5c54330245645f2ed40d56c74538902#diff-81f758fb88f738604ba750f9dde1f6bbL9